fixed ambulance delete bug

+++ app/src/main/java/com/swasthavyas/emergencyllp/component/dashboard/owner/ambulance/worker/DeleteAmbulanceWorker.java
@@ -6,10 +6,13 @@ import android.util.Log;
 import androidx.annotation.NonNull;
 import androidx.work.WorkerParameters;
 
+import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.firestore.WriteBatch;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
 import com.swasthavyas.emergencyllp.component.dashboard.owner.ambulance.domain.model.Ambulance;
+import com.swasthavyas.emergencyllp.component.dashboard.owner.employee.domain.model.EmployeeDriver;
 import com.swasthavyas.emergencyllp.util.AppConstants;
 import com.swasthavyas.emergencyllp.util.asyncwork.ListenableWorkerAdapter;
 import com.swasthavyas.emergencyllp.util.asyncwork.NetworkResultCallback;
@@ -26,8 +29,9 @@ public class DeleteAmbulanceWorker extends ListenableWorkerAdapter {
         String ambulanceId = getInputData().getString(Ambulance.ModelColumns.ID);
         String ownerId = getInputData().getString(Ambulance.ModelColumns.OWNER_ID);
         String imageRef = getInputData().getString(Ambulance.ModelColumns.IMAGE_REF);
+        String vehicleNumber = getInputData().getString(Ambulance.ModelColumns.VEHICLE_NUMBER);
 
-        if(ambulanceId == null || ownerId == null || imageRef == null) {
+        if(ambulanceId == null || ownerId == null || imageRef == null || vehicleNumber == null) {
             Log.d(AppConstants.TAG, "doAsyncBackgroundTask: " + getInputData());
             callback.onFailure(new NullPointerException("ambulanceId or ownerId or imageRef is null"));
             return;
@@ -38,29 +42,48 @@ public class DeleteAmbulanceWorker extends ListenableWorkerAdapter {
         dbInstance
                 .collection("owners")
                 .document(ownerId)
-                .collection("ambulances")
-                .document(ambulanceId)
-                .delete()
-                .addOnCompleteListener(deleteTask -> {
-                    if(deleteTask.isSuccessful()) {
+                .collection("employees")
+                .whereEqualTo(EmployeeDriver.ModelColumns.ASSIGNED_AMBULANCE_NUMBER, vehicleNumber)
+                .get()
+                .addOnCompleteListener(employeeFetchTask -> {
+                   if(employeeFetchTask.isSuccessful()) {
+                       WriteBatch updateBatch = dbInstance.batch();
+                       for(DocumentSnapshot snapshot : employeeFetchTask.getResult()) {
+                           if(snapshot.exists()) {
+                               updateBatch.update(snapshot.getReference(), EmployeeDriver.ModelColumns.ASSIGNED_AMBULANCE_NUMBER, "None");
+                           }
+                       }
 
-                        FirebaseStorage storage = FirebaseStorage.getInstance();
+                       updateBatch.commit();
 
-                        StorageReference ambulanceRef = storage.getReferenceFromUrl(imageRef);
+                       dbInstance
+                               .collection("owners")
+                               .document(ownerId)
+                               .collection("ambulances")
+                               .document(ambulanceId)
+                               .delete()
+                               .addOnCompleteListener(deleteTask -> {
+                                   if(deleteTask.isSuccessful()) {
 
-                        ambulanceRef.delete()
-                                        .addOnCompleteListener(deleteImageTask -> {
-                                            if(deleteImageTask.isSuccessful()) {
-                                                callback.onSuccess(null);
-                                            }
-                                            else {
-                                                callback.onFailure(deleteImageTask.getException());
-                                            }
-                                        });
-                    }
-                    else {
-                        callback.onFailure(deleteTask.getException());
-                    }
+                                       FirebaseStorage storage = FirebaseStorage.getInstance();
+
+                                       StorageReference ambulanceRef = storage.getReferenceFromUrl(imageRef);
+
+                                       ambulanceRef.delete()
+                                               .addOnCompleteListener(deleteImageTask -> {
+                                                   if(deleteImageTask.isSuccessful()) {
+                                                       callback.onSuccess(null);
+                                                   }
+                                                   else {
+                                                       callback.onFailure(deleteImageTask.getException());
+                                                   }
+                                               });
+                                   }
+                                   else {
+                                       callback.onFailure(deleteTask.getException());
+                                   }
+                               });
+                   }
                 });
 
     }
